requirements.txt
openai
python-dotenv
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
requests